all repos were added

+++ src/main/java/com/github/cassette/model/repositories/IMessageStatusRepository.java
@@ -5,5 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface IMessageStatusRepository extends JpaRepository<MessageStatusEntity, Integer> {
+public interface IMessageStatusRepository extends JpaRepository<MessageStatusEntity, Long> {
+    
 }
